Ch. 17 §02: correct numbering and content for Listing 17-07
