+++ listings/ch17-async-await/listing-17-07-orig/src/main.rs
@@ -1,0 +1,45 @@
+use std::time::Duration;
+
+fn main() {
+    trpl::block_on(async {
+        let (tx, mut rx) = trpl::channel();
+
+        let tx_fut = async {
+            let vals = vec![
+                String::from("hi"),
+                String::from("from"),
+                String::from("the"),
+                String::from("future"),
+            ];
+
+            for val in vals {
+                tx.send(val).unwrap();
+                trpl::sleep(Duration::from_secs(1)).await;
+            }
+        };
+
+        let rx_fut = async {
+            while let Some(received) = rx.recv().await {
+                println!("Got: {received}");
+            }
+        };
+
+        // ANCHOR: updated
+        let tx_fut2 = async {
+            let vals = vec![
+                String::from("more"),
+                String::from("messages"),
+                String::from("for"),
+                String::from("you"),
+            ];
+
+            for val in vals {
+                tx.send(val).unwrap();
+                trpl::sleep(Duration::from_secs(1)).await;
+            }
+        };
+
+        trpl::join3(tx_fut, tx_fut2, rx_fut).await;
+        // ANCHOR_END: updated
+    });
+}
